Updated the way fraction::reduce() works to eliminate skipped factors
Previously, if a fraction with a factor larger than the largest known prime was attempted to be reduced, it wouldn't reduce it properly, and would return the fraction without the factor removed.  Now, if reduce() encounters a fraction with a numerator & denominator greater than the largest known prime it will check every value from the last prime + 1 up to the numerator or denominator (whichever's lower) to see if they are factors.